+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## [Unreleased]
 ### Added
 - Autocompletion for Fusion paths in EEL
+### Removed
+- Support for TYPO3.Neos namespace
+- Support for .ts2 file extension
 
 ## [1.12.2]
 - Compatibility with 2022.1